mcp/requirements.txt
mcp
fastmcp
opencv-python-headless
pillow
numpy
google-genai
pytest-cov